requirements.txt
numpy
scipy
pandas
matplotlib
pillow
seaborn
joblib
shapely